refactor(monitor): convert page to server component and add client MonitorDetailsComponent
